requirements.txt
# Streamlit frontend
streamlit
python-dotenv

# RSS & Guardian scraping
feedparser
requests

# HTML parsing
beautifulsoup4           # provides `from bs4 import BeautifulSoup`  [oai_citation_attribution:0‡GitHub](https://github.com/google-research/bleurt/blob/master/bleurt/wmt/evaluator.py?utm_source=chatgpt.com)
readability-lxml

# Full‑text extraction (optional paths)
newspaper3k
trafilatura

# NER & summarization
transformers
torch
google-generativeai
tf-keras

# CPU‑only evaluation metrics
evaluate

# BLEURT metric and its TF dependencies
bluert git+https://github.com/google-research/bleurt.git    # installs the BLEURT package  [oai_citation_attribution:1‡GitHub](https://github.com/google-research/bleurt?utm_source=chatgpt.com)
tensorflow>=1.15                                     # BLEURT requires TF ≥1.15  [oai_citation_attribution:2‡GitHub](https://github.com/google-research/bleurt/blob/master/setup.py?utm_source=chatgpt.com)
tf-slim>=1.1                                         # BLEURT’s setup.py lists tf-slim≥1.1  [oai_citation_attribution:3‡GitHub](https://github.com/google-research/bleurt/blob/master/setup.py?utm_source=chatgpt.com)
sentencepiece                                        # BLEURT’s setup.py lists sentencepiece  [oai_citation_attribution:4‡GitHub](https://github.com/google-research/bleurt/blob/master/setup.py?utm_source=chatgpt.com)